% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cv.pls.R
\name{cv.pls}
\alias{cv.pls}
\title{Cross Validation for  Partial Least Squares Regression}
\usage{
cv.pls(x, y, random = TRUE, nfolds = 5, maxcomp = NULL,
  use.scale = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{x}{predictor matrix}

\item{y}{response vector}

\item{random}{randomization of y}

\item{nfolds}{number of folds - default is \code{5}.}

\item{maxcomp}{Maximum number of components included within the models,
if not specified, default is the variable (column) numbers in x.}

\item{use.scale}{scale}

\item{verbose}{shall we print the cross validation process}

\item{...}{other arguments that can be passed to \code{enpls.en}}
}
\value{
A list containing four components:
\itemize{
\item \code{y} - response vector
\item \code{ypred} - predicted y
\item \code{residual} - cross validation result (y.pred - y.real)
\item \code{RMSE} - RMSE
\item \code{R2} - R2
}
}
\description{
This function performs k-fold cross validation for
partial least squares regression.
}
\details{
This function performs k-fold cross validation for
partial least squares regression.
}
\examples{
data(alkanes)
x = alkanes$x
y = alkanes$y

set.seed(42)
cv.enpls.fit = cv.enpls(x, y, MCtimes = 10)
print(cv.enpls.fit)
plot(cv.enpls.fit)
}
\author{
Min-feng Zhu <\email{wind2zhu@163.com}>,
        Nan Xiao <\email{road2stat@gmail.com}>
}
\seealso{
See \code{\link{cv.enpls}} for Cross Validation for ensemble PLS regression.
}

