% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/enpls.ad.R
\name{enpls.ad.core}
\alias{enpls.ad.core}
\title{core function for enpls.od}
\usage{
enpls.ad.core(plsdf.x, plsdf.y, x.test, maxcomp = NULL)
}
\value{
predicted y for test set
}
\description{
select the best ncomp with cross-validation and
use it to fit the complete training set again,
then predict on the test set. scale = TRUE
}
\keyword{internal}

