% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/enpls.ad.R
\name{enpls.ad.cv}
\alias{enpls.ad.cv}
\title{cv function for enpls.od}
\usage{
enpls.ad.cv(plsdf.x, plsdf.y, nfolds = 5L, maxcomp = NULL)
}
\value{
5-fold cross validation result predicted y
}
\description{
This function performs k-fold cross validation for
ensemble partial least squares regression.
}
\keyword{internal}

