% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/enpls.ad.R
\name{enpls.ad}
\alias{enpls.ad}
\title{Ensemble Partial Least Squares for Applicability Domain}
\usage{
enpls.ad(x, y, x.test = NULL, y.test = NULL, maxcomp = NULL,
  MCtimes = 500L, verbose = FALSE, method = c("mc", "bootstrap",
  "jagging"), ratio = 0.8, parallel = 1L)
}
\arguments{
\item{x}{predictor matrix.}

\item{y}{response vector.}

\item{x.test}{predictor matrix for test.}

\item{y.test}{response vector for test.}

\item{maxcomp}{Maximum number of components included within the models,
if not specified, default is the variable (column) numbers in x.}

\item{MCtimes}{times of Monte-Carlo.}

\item{verbose}{shall we print the MCtimes process.}

\item{method}{\code{"mc"} or \code{"bootstrap"} or \code{"jagging"}. Default is \code{"mc"}.}

\item{ratio}{sample ratio used when \code{method = "mc"} and \code{method = "jagging"}.}

\item{parallel}{Integer. Number of parallel processes to use. Default is \code{1}, which means run serially.}
}
\value{
A list containing four components:
\itemize{
\item \code{STD.cv} - STD value for training set
\item \code{STD.te} - STD value for test set
\item \code{error.cv} - absolute prediction error of training set
\item \code{error.te} - absolute prediction error of test set
}
}
\description{
This function performs applicability domain with ensemble partial least squares.
}
\details{
This function performs applicability domain with ensemble partial least squares.
}
\examples{
data(logS)
x = logS$x
y = logS$y
x.test1 = logS$x.test1
y.test1 = logS$y.test1

set.seed(42)
ad_test1 = enpls.ad(x, y, x.test1, y.test1, MCtimes = 10)
print(ad_test1)
plot(ad_test1)

x.test2 = logS$x.test2
y.test2 = logS$y.test2

set.seed(42)
ad_test2 = enpls.ad(x, y, x.test2, y.test2, MCtimes = 10)
print(ad_test2)
plot(ad_test2)
}
\author{
Min-feng Zhu <\email{wind2zhu@163.com}>,
        Nan Xiao <\email{road2stat@gmail.com}>
}
\references{
Kaneko H, Funatsu K.
"Applicability Domain Based on Ensemble Learning in Classification and Regression Analyses."
\emph{Journal of chemical information and modeling} 54, no. 9 (2014): 2469-2482.
}
\seealso{
See \code{\link{enpls.fs}} for feature selection with ensemble PLS.
See \code{\link{enpls.en}} for ensemble PLS regression.
See \code{\link{enpls.od}} for Outlier Detection with ensemble PLS
}

