% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/enplscv.R
\name{enplscv}
\alias{enplscv}
\title{Ensemble Partial Least Squares Cross-Valdidation for feature selection}
\usage{
enplscv(trainx, trainy, cv.fold = 5, scale = "log", step = 0.5,
  recursive = FALSE, MCtimes = 10, ...)
}
\arguments{
\item{trainx}{matrix or data frame containing columns of predictor}

\item{trainy}{vector of response, must have length equal to the number
of rows in \code{trainx}.}

\item{cv.fold}{number of folds in the cross-validation}

\item{scale}{if \code{"log"}, reduce a fixed proportion (\code{step})
of variables at each step, otherwise reduce \code{step} variables at a
time.}

\item{step}{if \code{log=TRUE}, the fraction of variables to remove at
each step, else remove this many variables at a time}

\item{recursive}{whether variable importance is (re-)assessed at each
step of variable reduction}

\item{MCtimes}{times of Monte-Carlo}

\item{...}{other arguments passed on to \code{enpls.en}}
}
\value{
A list with the following components:
\itemize{
\item \code{n.var} - vector of number of variables used at each step.
\item \code{error.cv} - corresponding vector of error rates or MSEs at each step.
\item \code{predicted} - list of \code{n.var} components, each containing
the predicted values from the cross-validation.
\item \code{res} - list of \code{n.var} components, each containing
the sum of coefficient values from the cross-validation.
\item \code{imp} - list of \code{n.var} components, each containing
the sum of coefficient values from the cross-validation.
}
}
\description{
This function shows the cross-validated prediction performance of
models with sequentially reduced number of predictors (ranked by
variable importance) via a nested cross-validation procedure.
}
\details{
This function shows the cross-validated prediction performance of
models with sequentially reduced number of predictors (ranked by
variable importance) via a nested cross-validation procedure.
}
\examples{
data(logS)
x = logS$x
y = logS$y

result = enplscv(x, y, recursive = TRUE)
with(result, plot(n.var, error.cv, type="b", lwd=2))
}
\author{
Min-feng Zhu <\email{wind2zhu@163.com}>,
        Nan Xiao <\email{road2stat@gmail.com}>
}
\references{
Svetnik, V., Liaw, A., Tong, C. and Wang, T., "Application of Breiman's
Random Forest to Modeling Structure-Activity Relationships of
Pharmaceutical Molecules", MCS 2004, Roli, F. and Windeatt, T. (Eds.)
pp. 334-343.
}
\seealso{
See \code{\link{enpls.en}} for feature selection with ensemble PLS.
See \code{\link{enpls.en}} for ensemble PLS regression
}

