% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/enpls-datalist.R
\docType{data}
\name{logS}
\alias{logS}
\title{The aqueous solubility data}
\format{A list with 6 components:
\itemize{
\item x - data frame with 549 rows (samples) and 166 columns (predictors)
\item y - numeric vector of length 549 (response)
\item x.test1 - data frame with 274 rows (samples) and 166 columns (predictors)
\item y.test1 - numeric vector of length 274 (response)
\item x.test2 - data frame with 446 rows (samples) and 166 columns (predictors)
\item y.test2 - numeric vector of length 446 (response)
}}
\usage{
data(logS)
}
\description{
The aqueous solubility data reported by Hou et al.
}
\details{
This dataset contains 549 aqueous solubility samples (y) which have been modeled
by 166 molecular descriptors (x) for training set, 274 aqueous solubility samples
(y) which have been modeled by 166 molecular descriptors (x) for test1 set and 446
aqueous solubility samples (y) which have been modeled by 166 molecular descriptors
(x) for test2 set.

Molecular descriptor types:
\itemize{
\item Constitution
\item Topology
\item Connectivity
\item Kappa
\item EState
\item Autocorrelation-moran
\item Autocorrelation-geary
\item Autocorrelation-broto
\item Molecular properties
\item Charge
\item Moe-Type descriptors
}
}
\examples{
data(logS)
str(logS)
}
\references{
Hou, T. J.; Xia, K.; Zhang, W.; Xu, X. J.
"ADME Evaluation in Drug Discovery. 4. Prediction of Aqueous Solubility
Based on Atom Contribution Approach"
Journal of chemical information and computer sciences 44, no. 1 (2004): 266-275.
}

