% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.enpls.ad}
\alias{plot.enpls.ad}
\title{Plot enpls.ad Object}
\usage{
\method{plot}{enpls.ad}(x, main = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{enpls.ad}.}

\item{main}{plot title}

\item{...}{Other graphical parameters to be passed on to \code{ggplot}.}
}
\description{
This function plots enpls.ad object.
}
\details{
This function plots enpls.ad object.
}
\examples{
data(logS)
x = logS$x
y = logS$y
x.test = logS$x.test1
y.test = logS$y.test1

set.seed(42)
ad = enpls.ad(x, y, x.test, y.test, MCtimes = 10)
plot(ad)
}
\author{
Min-feng Zhu <\email{wind2zhu@163.com}>,
        Nan Xiao <\email{road2stat@gmail.com}>
}
\seealso{
See \code{\link{enpls.ad}} for applicability domain with ensemble PLS.
}

