% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.R
\name{print.enpls.en}
\alias{print.enpls.en}
\title{Print Fitted Ensemble Partial Least Squares Object}
\usage{
\method{print}{enpls.en}(x, ...)
}
\arguments{
\item{x}{An object of class \code{enpls.en}.}

\item{...}{Other parameters to be passed on to \code{print}.}
}
\description{
This function prints coefficients of each model in the enpls.en object.
}
\details{
This function prints coefficients of each model in the enpls.en object.
}
\examples{
data(alkanes)
x = alkanes$x
y = alkanes$y

set.seed(42)
enpls.fit = enpls.en(x, y, MCtimes = 10)
print(enpls.fit)
}
\author{
Min-feng Zhu <\email{wind2zhu@163.com}>,
        Nan Xiao <\email{road2stat@gmail.com}>
}
\seealso{
See \code{\link{enpls.en}} for performing ensemble PLS regression.
}

